# Alternate implementation of Cache