--- a/lib/test/t_Cache_std.expout
+++ b/lib/test/t_Cache_std.expout
@@ -1,10 +1,10 @@
 myCache = class=Cache<PersistentCollection<Scalar>, PersistentCollection<Scalar>> enabled=true name=aCache maxSize=3 size=0 hits=0 points={}
-myCache = class=Cache<PersistentCollection<Scalar>, PersistentCollection<Scalar>> enabled=true name=aCache maxSize=3 size=1 hits=0 points={[1,2,3]->[10,20]/0}
+myCache = class=Cache<PersistentCollection<Scalar>, PersistentCollection<Scalar>> enabled=true name=aCache maxSize=3 size=1 hits=0 points={([1,2,3],[10,20])}
 Is ko in myCache ? false
 Cache value for ko = []
 Is ok in myCache ? true
 Cache value for ok = [10,20]
-myCache = class=Cache<PersistentCollection<Scalar>, PersistentCollection<Scalar>> enabled=true name=aCache maxSize=3 size=1 hits=1 points={[1,2,3]->[10,20]/1}
-myCache = class=Cache<PersistentCollection<Scalar>, PersistentCollection<Scalar>> enabled=true name=aCache maxSize=3 size=2 hits=1 points={[1,2,3]->[10,20]/1, [2,4,6]->[20,40]/0}
-myCache = class=Cache<PersistentCollection<Scalar>, PersistentCollection<Scalar>> enabled=true name=aCache maxSize=3 size=3 hits=1 points={[1,2,3]->[10,20]/1, [2,4,6]->[20,40]/0, [3,6,9]->[30,60]/0}
-myCache = class=Cache<PersistentCollection<Scalar>, PersistentCollection<Scalar>> enabled=true name=aCache maxSize=3 size=3 hits=1 points={[1,2,3]->[10,20]/1, [3,6,9]->[30,60]/0, [4,8,12]->[40,80]/0}
+myCache = class=Cache<PersistentCollection<Scalar>, PersistentCollection<Scalar>> enabled=true name=aCache maxSize=3 size=1 hits=1 points={([1,2,3],[10,20])}
+myCache = class=Cache<PersistentCollection<Scalar>, PersistentCollection<Scalar>> enabled=true name=aCache maxSize=3 size=2 hits=1 points={([2,4,6],[20,40]), ([1,2,3],[10,20])}
+myCache = class=Cache<PersistentCollection<Scalar>, PersistentCollection<Scalar>> enabled=true name=aCache maxSize=3 size=3 hits=1 points={([3,6,9],[30,60]), ([2,4,6],[20,40]), ([1,2,3],[10,20])}
+myCache = class=Cache<PersistentCollection<Scalar>, PersistentCollection<Scalar>> enabled=true name=aCache maxSize=3 size=3 hits=1 points={([4,8,12],[40,80]), ([3,6,9],[30,60]), ([2,4,6],[20,40])}