With the previous implementation, when the Cache is full, the least recently used element is searched by traversing all elements whenever a new element is inserted.

This new implementation is more standard, it stores keys in a doubly-linked list, and uses a map to store the position of each key in this iterator. Key is moved at the beginning at the list when it is accessed, so tail list always contains the last recently used key.

Since the cache size is fixed, it would be better to use a hash map, but OT does not use hash map anywhere.