@dbarbier 2017-11-10
Here is a simple benchmark:
```
import openturns as ot
import time

dim = 20
size = 100000
cacheSize = 1024      # Default
cacheSize = size

ot.ResourceMap.SetAsUnsignedInteger("cache-max-size", cacheSize)
X = ot.Normal(dim).getSample(size)
Y = ot.Normal(dim).getSample(size)
tic = time.time()
f = ot.DatabaseFunction(X,Y)
print("Insertion: ", time.time() - tic)

s = ot.Point(dim)
tic = time.time()
for i in xrange(size):
  s += f(X[i])
s /= size
print("Access: ", time.time() - tic)
print(s)
print(Y.computeMean())
```

Tests have been run without the 2nd commit, to test element insertion when cache is full.

Old implementation:
 * size=100000 cacheSize=1024 Insertion=24.6s Access=1.86s
 * size=100000 cacheSize=size Insertion=0.26s Access=1.37s

New implementation:
 * size=100000 cacheSize=1024 Insertion=0.24s Access=1.88s
 * size=100000 cacheSize=size Insertion=0.28s Access=1.41s

These figures come from a single run in each case, it would be nice to have more figures, if you have ideas on real use cases.  It is pretty clear that insertion performs much better when cache is full (the 2nd commit makes this point moot when creating a DatabaseFunction, but there are other cases where elements are added into a full cache), but it is not clear whether access time is comparable or not.