@regislebrun 2017-11-10
@dbarbier The gain looks impressive, but I am not sure it will change our typical user's life (using MonteCarlo with continuous distributions). The cache is expected to be activated for costly functions for which the cache overhead is negligible and when a significant portion of the evaluations is expected to be repeated. It has been implemented to speed-up algorithms like finite difference, optimization in relation with FORM/SORM and other places I can't remember. Nevertheless, I may miss some situations where the cache is used even if a lot of redundant evaluations are expected, in which case your PR is a must.