@dbarbier 2017-11-10
@regislebrun DatabaseEvaluation::setSample contains this comment:
> // Don't activate the cache systematically as it can take a significant amount of time for large samples
So I guess that this performance penalty had been noticed.