@dbarbier 2017-11-10
I am moving the 2nd commit into a separate PR for clarity.