@regislebrun 2017-11-11
@dbarbier For DatabaseEvaluation, IMO the cache should be deactivated for this class, as the KDTree search should be as efficient as the search in the cache. BTW, the evaluation operator based on a sample has no cache at all. But you are right, it has been noticed that the cache was a real pain for this specific function!